% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_binom_recur.r
\name{calc_binom_recur}
\alias{calc_binom_recur}
\title{Probability calc}
\usage{
calc_binom_recur(gene_sets)
}
\arguments{
\item{gene_sets}{A matrix of genes (rows)  by study (columns)}
}
\value{
\code{fdrs_bin}
}
\description{
Probability calc
}
\examples{

gene_sets <- matrix(  rbinom(1000,1,0.05), ncol=10, nrow=100)
fdrs_bin <- calc_binom_recur(gene_sets)

}
