% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DE.R
\name{calc_DE}
\alias{calc_DE}
\title{Run differential expression}
\usage{
calc_DE(counts, groups, method, row_filter = FALSE, col_filter = FALSE, ...)
}
\arguments{
\item{counts}{A matrix of counts}

\item{groups}{conditions}

\item{method}{DE method to use}

\item{row_filter}{filter for rows}

\item{col_filter}{filter for columns}

\item{...}{plotting parameters}
}
\value{
\code{combined_out}
}
\description{
Run differential expression
}
\examples{


#load("data/counts_data.Rdata")
counts <- matrix(  2 ^ rpois(1e5, 5) , ncol=10, nrow=1000)
rownames(counts) <- paste0('gene', 1:1000 )
colnames(counts) <- paste0('sample', 1:10 )

groups <- c( rep(1,5), rep(2,5) )
method <- 'wilcox'

degs <- calc_DE(counts, groups, method)

}
