% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fdrs_recur.r
\name{calc_fdrs_recur}
\alias{calc_fdrs_recur}
\title{Probability calc}
\usage{
calc_fdrs_recur(data, pp = 0.05, nr = 1000)
}
\arguments{
\item{data}{A matrix}

\item{pp}{significance/FDR threshold}

\item{nr}{number of repeats to run}
}
\value{
\code{res}
}
\description{
Probability calc
}
\examples{

genesets <- matrix(  rbinom(1000,1,0.05), ncol=10, nrow=100)
fdrs_recur <- calc_fdrs_recur(genesets)

}
