% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_functional_effects_network.R
\name{calculate_functional_effects_network}
\alias{calculate_functional_effects_network}
\title{Connectivity tests}
\usage{
calculate_functional_effects_network(
  genesets,
  network,
  nettype,
  dir,
  studies,
  nsub,
  nr = 1000
)
}
\arguments{
\item{genesets}{Sets of genes to test}

\item{network}{A matrix (network)}

\item{nettype}{network label}

\item{dir}{output directory}

\item{studies}{gene set labels}

\item{nsub}{minimum gene set size of studies being compared (downsample to this)}

\item{nr}{number of repeats}
}
\value{
\code{output} list with effect sizes for network connectivities
}
\description{
Connectivity tests
}
\examples{

genesets <- matrix(  rbinom(1000,1,0.1), ncol=10, nrow=100)
studies <- paste0('study', 1:10 )
genes   <-  paste0('gene', 1:100 )
colnames(genesets) <- studies
rownames(genesets) <- genes

network <- diag(1000)
upper <- row(network) < col(network)
network[upper] <- runif(sum(upper), 0,1 )
network <- network + t(network)
diag(network) <- 1
rownames(network) <-paste0('gene', 1:1000 )
colnames(network) <-paste0('gene', 1:1000 )
nettype <- 'random'
dir <- 'out_test'


}
