% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_coexp.R
\name{cluster_coexp}
\alias{cluster_coexp}
\title{Cluster sub-network}
\usage{
cluster_coexp(
  coexp,
  runid = "",
  filt_min = 6,
  medK = 0.5,
  col_map = viridis(100),
  method = "average",
  flag_plot = FALSE,
  flag_med = TRUE,
  flag_dist = FALSE,
  frac = 0.995,
  deep_split = 2,
  min_cs = 2
)
}
\arguments{
\item{coexp}{A matrix (co-expression or other network)}

\item{runid}{ID title}

\item{filt_min}{minimum size of cluster to consider}

\item{medK}{median expression value of co-expression network}

\item{col_map}{color palette for heatmaps}

\item{flag_plot}{Boolean}
}
\value{
\code{clusterids}
}
\description{
Cluster sub-network
}
\examples{

cpm <- matrix(  abs(rnorm(100)*10) , ncol=1, nrow=100)
cpm <- sapply(1:10, function(i) cpm[,1])
cpm <- jitter(cpm, amount = 4)
network <- matrix(rank(cor(t(cpm))), nrow=100) - 1
network <- network/max(network)
diag(network) <- 1
rownames(network) <-paste0('gene', 1:100 )
colnames(network) <-paste0('gene', 1:100 )
nettype <- 'random'
dir <- 'out_test'
cluster_coexp(network)

}
