% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_recur.r
\name{count_recur}
\alias{count_recur}
\title{Recurrence count}
\usage{
count_recur(data, nmax)
}
\arguments{
\item{data}{A matrix}

\item{nmax}{number of datasets}
}
\value{
\code{res}
}
\description{
Recurrence count
}
\examples{

data <-  rowSums(matrix(  rbinom(1000,1,0.05), ncol=10, nrow=100))
count_recur(as.matrix(data), 10)

}
