% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_data.R, R/genesets_down.R,
%   R/genesets_up.R, R/res.brain.r
\docType{data}
\name{counts_data}
\alias{counts_data}
\title{count data from TAF1 cohort}
\format{A matrix with 60K rows and 26 columns
\describe{
  \item{chr}{chromosome}

}
 @source \url{}}
\description{
A dataset containing the expression level counts

A dataset containing the expression level counts

A dataset containing the expression level counts

A dataset containing the expression level counts
}
