% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_network_hdf5.R
\name{download_network_hdf5}
\alias{download_network_hdf5}
\title{Download networks from dropbox}
\usage{
download_network_hdf5(network_type = "generic", flag_occr = TRUE, dir = "")
}
\arguments{
\item{network_type}{name of pre-built network to use (generic, blood, brain)}

\item{flag_occr}{boolean to select occurence network or coexpression network}

\item{dir}{dummy directory for the networks}
}
\description{
Download networks from dropbox
}
\examples{
network_type='generic'

}
