% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_sub_network_list_random.R
\name{explore_sub_network_list_random}
\alias{explore_sub_network_list_random}
\title{Connectivity tests}
\usage{
explore_sub_network_list_random(network, dir, label, ns, nr = 1000)
}
\arguments{
\item{network}{A matrix (network)}

\item{dir}{output directory}

\item{label}{name}

\item{ns}{number of genes to subsample to}

\item{nr}{number of repeats}
}
\value{
\code{data}
}
\description{
Connectivity tests
}
\examples{

genesets <- matrix(  rbinom(1000,1,0.1), ncol=10, nrow=100)
studies <- paste0('study', 1:10 )
studies_genes   <-  paste0('gene', 1:100 )
colnames(genesets) <- studies
rownames(genesets) <- studies_genes
i <- 1
gene_list <- as.matrix(cbind(studies_genes, as.numeric(genesets[,i]))[genesets[, i] > 0, ])[,1]
ns <- length(gene_list)

network <- diag(1000)
upper <- row(network) < col(network)
network[upper] <- runif(sum(upper), 0,1 )
network <- network + t(network)
diag(network) <- 1
rownames(network) <-paste0('gene', 1:1000 )
colnames(network) <-paste0('gene', 1:1000 )
nettype <- 'random'
dir <- '.'

# explore_sub_network_list_random(network, 'output', 'test', ns)


}
