% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_enrichment.R
\name{gene_set_enrichment}
\alias{gene_set_enrichment}
\title{Gene set enrichment}
\usage{
gene_set_enrichment(genes, genes.labels, voc)
}
\arguments{
\item{genes}{an array of numeric values}

\item{genes.labels}{GO matrix or some such}

\item{voc}{term descriptions}
}
\value{
\code{results}
}
\description{
Gene set enrichment
}
\examples{
data <-  rnorm(1000)
geo_mean(data)

}
