% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_enrichment.R
\name{gene_set_enrichment_aucs}
\alias{gene_set_enrichment_aucs}
\title{Gene set enrichment - AUC version}
\usage{
gene_set_enrichment_aucs(gene_sets, gene_rankings)
}
\arguments{
\item{gene_sets}{GO matrix or some such}

\item{gene_rankings}{an array of numeric values}
}
\value{
\code{results}
}
\description{
Gene set enrichment - AUC version
}
