% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{geo_mean}
\alias{geo_mean}
\title{OutDeCo helper functions}
\usage{
geo_mean(data)
}
\arguments{
\item{data}{an array of numeric values}
}
\value{
\code{gm}
}
\description{
geometric mean
}
\examples{
data <-  rnorm(1000)
geo_mean(data)

}
