% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{geo_sd}
\alias{geo_sd}
\title{geometric standard deviation}
\usage{
geo_sd(data)
}
\arguments{
\item{data}{an array of numeric values}
}
\value{
\code{gs}
}
\description{
geometric standard deviation
}
\examples{
data <-  rnorm(1000)
geo_sd(data)

}
