% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_se.R
\name{geo_se}
\alias{geo_se}
\title{geometric standard error}
\usage{
geo_se(data)
}
\arguments{
\item{data}{an array of numeric values}
}
\value{
\code{gse}
}
\description{
geometric standard error
}
\examples{

data <-  rnorm(1000)
geo_se(data)

}
