% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recur_mat.R
\name{get_recur_mat}
\alias{get_recur_mat}
\title{Recurrence to frequency matrix}
\usage{
get_recur_mat(recurs)
}
\arguments{
\item{recurs}{A vector of gene recurrences}
}
\value{
\code{mat10}
}
\description{
Recurrence to frequency matrix
}
\examples{

data_pre <- as.matrix(rowSums(matrix(  rbinom(1000,1,0.1), ncol=10, nrow=100)))
data_post <- data_pre - (data_pre * rbinom(100, 1, 0.2 ))
recurs = cbind(data_post, data_pre)

get_recur_mat(recurs)

}
