% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_studies.R
\docType{data}
\name{pd_studies}
\alias{pd_studies}
\title{Metadata}
\format{
A matrix with 15 rows and 9 columns
\describe{
  \item{study_names}
  \item{Tissue}
  \item{Cases}
  \item{PMID}
  \item{file}
  \item{dataset}
  \item{label}
  \item{Method}
  \item{Platform}
}
 @source \url{}
}
\description{
Metadata for Parkinson's disease studies used in the meta-analysis
}
