% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_hist.R
\name{plot_2D_hist}
\alias{plot_2D_hist}
\title{Plot 2D histogram}
\usage{
plot_2D_hist(mat, pt.x, pt.y, ...)
}
\arguments{
\item{mat}{yA matrix of frequencies}

\item{pt.x}{threshold x}

\item{pt.y}{threshold}

\item{...}{parameters to pass onto the image() function}
}
\description{
Plot 2D histogram
}
\examples{

data_pre <- as.matrix(rowSums(matrix(  rbinom(1000,1,0.1), ncol=10, nrow=100)))
data_post <- data_pre - (data_pre * rbinom(100, 1, 0.2 ))
recurs = cbind(data_post, data_pre)

mat10 <- get_recur_mat(recurs)
plot_2D_hist(mat10, 1, 4)

}
