% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coexpression_heatmap.R
\name{plot_coexpression_heatmap}
\alias{plot_coexpression_heatmap}
\title{Plot coexpression heatmap}
\usage{
plot_coexpression_heatmap(
  coexp,
  cluster_output,
  col_map = viridis(100),
  filt = FALSE,
  filt_min = 6
)
}
\arguments{
\item{coexp}{coexpression matrix}

\item{cluster_output}{list}

\item{col_map}{color palette (array)}

\item{filt}{boolean to show filtered genes based on filtMin}

\item{filt_min}{minimum cluster size to mark for filtering}
}
\description{
Plot coexpression heatmap
}
