% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_networks.R
\name{plot_compare_networks}
\alias{plot_compare_networks}
\title{A plot of two networks (and their joining river)}
\usage{
plot_compare_networks(net1, net2, clust_net1, clust_net2)
}
\arguments{
\item{net1}{first network}

\item{net2}{second network}

\item{clust_net1}{clusters of first}

\item{clust_net2}{clusters of second}
}
\description{
A plot of two networks (and their joining river)
}
