% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot recurrence histogram}
\usage{
plot_enrichment(data, n_max, ...)
}
\arguments{
\item{data}{values}

\item{n_max}{total number of studies}
}
\description{
Plot recurrence histogram
}
\examples{

data <-  matrix(  rbinom(1000,1,0.05), ncol=10, nrow=100)
n_max <- dim(data)[2]
plot_enrichment( data, n_max)

}
