% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_gene_set_enrichment}
\alias{plot_gene_set_enrichment}
\title{Plot gene set enrichment results}
\usage{
plot_gene_set_enrichment(data, gene_list, gene_sets)
}
\arguments{
\item{data}{gene set enrichment results output}

\item{gene_list}{list of genes used as input into the gene set enrichment}

\item{gene_sets}{gene sets used as input into the gene set enrichment analysis}
}
\description{
Plot gene set enrichment results
}
