% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot network}
\usage{
plot_network(sub_net, clust_net, threshold = 0.5, filt_min = 6)
}
\arguments{
\item{sub_net}{network of genes}

\item{clust_net}{output from cluster_coexp}

\item{threshold}{weighted edge threshold to plot}

\item{filt_min}{minimum cluster size to display in plot}
}
\description{
Plot network
}
\examples{

network <- diag(100)
upper <- row(network) < col(network)
network[upper] <- runif(sum(upper), 0,1 )
network <- network + t(network)
diag(network) <- 1
rownames(network) <-paste0('gene', 1:100 )
colnames(network) <-paste0('gene', 1:100 )
clust_net <- cluster_coexp(network)
plot_network(network, clust_net)

}
