% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recurrence.r
\name{plot_recurrence}
\alias{plot_recurrence}
\title{Plot recurrence histogram}
\usage{
plot_recurrence(data, fdrs, n_max, flag_plot = "hist", ...)
}
\arguments{
\item{fdrs}{output from calc_fdrs_recur}

\item{n_max}{total number of studies}

\item{flag_plot}{plot type}

\item{recur}{values}
}
\description{
Plot recurrence histogram
}
\examples{

data <-  matrix(  rbinom(1000,1,0.05), ncol=10, nrow=100)
n_max <- dim(data)[2]
fdrs <- calc_fdrs_recur( data )
recur <- rowSums(data, na.rm=TRUE)
plot_recurrence( data, fdrs, n_max)

}
