% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_hist.R
\name{plot_scatter_hist}
\alias{plot_scatter_hist}
\title{Plot scatterplot and histograms}
\usage{
plot_scatter_hist(
  x,
  y,
  xlab = "",
  ylab = "",
  main = "",
  xybreaks = 100,
  clusters = FALSE,
  ...
)
}
\arguments{
\item{x}{values}

\item{y}{values}

\item{xlab}{x axis labe}

\item{ylab}{y axis label}

\item{main}{main title}

\item{xybreaks}{number of breaks for histogram}

\item{clusters}{output from cluster_coexp}

\item{...}{plotting parameters}
}
\description{
Plot scatterplot and histograms
}
\examples{
#'
x <- (rnorm(1000))
y <- (jitter(x)) + (rnorm(1000))
plot_scatter_hist(x,y)

}
