% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_degs.R
\name{reformat_degs}
\alias{reformat_degs}
\title{Reformat DE output}
\usage{
reformat_degs(output, method = "wilcox")
}
\arguments{
\item{output}{DE output}

\item{method}{method used}
}
\value{
\code{combined_out}
}
\description{
Reformat DE output
}
\examples{

}
