% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_connectivity_score_sub.R
\name{residual_connectivity_score_sub}
\alias{residual_connectivity_score_sub}
\title{Connectivity tests}
\usage{
residual_connectivity_score_sub(
  network,
  dir,
  label,
  gene_list,
  n_sub,
  n_r,
  fileout
)
}
\arguments{
\item{network}{A matrix (network)}

\item{dir}{output directory}

\item{label}{name}

\item{gene_list}{gene set of interest}

\item{n_sub}{number of genes to subsample}

\item{n_r}{number of repeats}

\item{fileout}{filename}
}
\description{
Connectivity tests
}
\examples{

genesets <- matrix(  rbinom(1000,1,0.1), ncol=10, nrow=100)
studies <- paste0('study', 1:10 )
studies_genes   <-  paste0('gene', 1:100 )
colnames(genesets) <- studies
rownames(genesets) <- studies_genes
i <- 3
gene_list <- as.matrix(cbind(studies_genes,
               as.numeric(genesets[,i]))[genesets[, i] > 0, ])[,1]

network <- diag(1000)
upper <- row(network) < col(network)
network[upper] <- runif(sum(upper), 0,1 )
network <- network + t(network)
diag(network) <- 1
rownames(network) <- paste0('gene', 1:1000 )
colnames(network) <- paste0('gene', 1:1000 )
nettype <- 'random'
dir <- 'out_test'
label <- studies[i]
fileout <- 'output'

#residual_connectivity_score_sub(network, dir, label,
#gene_list, 20, 100, fileout)

}
