% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\title{residuals from identity line}
\usage{
residuals(x, y, a, b, c)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{a}{maximum x value}

\item{b}{maximum y value}

\item{c}{y intercept}
}
\value{
\code{res}
}
\description{
residuals from identity line
}
\examples{
x <- (rnorm(1000))
y <- (jitter(x)) + (rnorm(1000))
rs <- residuals(x,y, -1,1,0)

plot(x,y, pch=19, cex=1.5)
abline(0,1, col='grey', lty=2, lwd=2)
segments(x, y - rs, x ,y , col=4)

}
