% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_filtering.R, R/run_filtering_entrez.R
\name{run_filtering}
\alias{run_filtering}
\title{Run the recurrence and filtering pipeline}
\usage{
run_filtering(
  subgenesets,
  flag,
  network_type,
  flag_out = FALSE,
  filt_min = 6,
  flag_med = TRUE,
  flag_dist = FALSE
)

run_filtering(
  subgenesets,
  flag,
  network_type,
  flag_out = FALSE,
  filt_min = 6,
  flag_med = TRUE,
  flag_dist = FALSE
)
}
\arguments{
\item{subgenesets}{DE genes}

\item{flag}{up or down regulated genes}

\item{flag_out}{boolean to save intermediate output files}

\item{net}{network to use}
}
\value{
\code{clusterids}

\code{clusterids}
}
\description{
Run the recurrence and filtering pipeline

Run the recurrence and filtering pipeline
}
\examples{

cpm <- matrix(  abs(rnorm(100)*10) , ncol=1, nrow=100)
cpm <- sapply(1:10, function(i) cpm[,1])
cpm <- jitter(cpm, amount = 4)
network <- matrix(rank(cor(t(cpm))), nrow=100) - 1
network <- network/max(network)
diag(network) <- 1
rownames(network) <-paste0('gene', 1:100 )
colnames(network) <-paste0('gene', 1:100 )
nettype <- 'random'
dir <- 'out_test'
cluster_coexp(network)


cpm <- matrix(  abs(rnorm(100)*10) , ncol=1, nrow=100)
cpm <- sapply(1:10, function(i) cpm[,1])
cpm <- jitter(cpm, amount = 4)
network <- matrix(rank(cor(t(cpm))), nrow=100) - 1
network <- network/max(network)
diag(network) <- 1
rownames(network) <-paste0('gene', 1:100 )
colnames(network) <-paste0('gene', 1:100 )
nettype <- 'random'
dir <- 'out_test'
cluster_coexp(network)

}
