% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{shuffle_cols}
\alias{shuffle_cols}
\title{Shuffle columns}
\usage{
shuffle_cols(data)
}
\arguments{
\item{data}{A matrix}
}
\value{
\code{output}
}
\description{
Shuffle columns
}
\examples{
data <- matrix(  rbinom(1000,1,0.05), ncol=10, nrow=100)
shuffle_cols(data)

}
