% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_rows.r
\name{shuffle_rows}
\alias{shuffle_rows}
\title{Shuffle rows}
\usage{
shuffle_rows(data)
}
\arguments{
\item{data}{A matrix}
}
\value{
\code{output}
}
\description{
Shuffle rows
}
\examples{

data <- matrix(  rbinom(1000,1,0.05), ncol=10, nrow=100)
shuffle_rows(data)


}
