% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_sample_resd.R
\name{sub_sample_resd}
\alias{sub_sample_resd}
\title{Connectivity test wrapper function}
\usage{
sub_sample_resd(net, dir, label, gene_list, n_sub, n_r, filename)
}
\arguments{
\item{net}{A matrix (network)}

\item{dir}{output directory}

\item{label}{name of study}

\item{gene_list}{Gene set to test}

\item{n_sub}{minimum gene set size of studies being compared}

\item{n_r}{number of repeats}

\item{filename}{fileanme}
}
\description{
Connectivity test wrapper function
}
\examples{

genesets <- matrix(  rbinom(1000,1,0.1), ncol=10, nrow=100)
studies <- paste0('study', 1:10 )
studies_genes   <-  paste0('gene', 1:100 )
colnames(genesets) <- studies
rownames(genesets) <- studies_genes
i <- 3
gene_list <- as.matrix(cbind(studies_genes,
             as.numeric(genesets[,i]))[genesets[, i] > 0, ])[,1]

network <- diag(1000)
upper <- row(network) < col(network)
network[upper] <- runif(sum(upper), 0,1 )
network <- network + t(network)
diag(network) <- 1
rownames(network) <-paste0('gene', 1:1000 )
colnames(network) <-paste0('gene', 1:1000 )

nettype <- 'random'
dir <- 'out_test'
label <- studies[i]
filename <- 'output'

# sub_sample_resd(network, dir, label, gene_list, n_sub, n_r, filename)

}
