% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{subset_network}
\alias{subset_network}
\title{Get sub-network}
\usage{
subset_network(
  deg,
  network,
  method = "top",
  n_top = 100,
  q = 0.05,
  fct = 2,
  min_cpm = 1
)
}
\arguments{
\item{deg}{A matrix (differential expression analysis)}

\item{network}{A matrix  (co-expression or other network)}

\item{method}{top (fc only) or reg (fc and q)}

\item{n_top}{number of genes to consider}

\item{q}{q-value}

\item{fct}{abs fold change threshold}
}
\value{
\code{combined_output}
}
\description{
Get sub-network
}
\examples{

network <- diag(1000)
upper <- row(network) < col(network)
network[upper] <- runif(sum(upper), 0,1 )
network <- network + t(network)
diag(network) <- 1
rownames(network) <-paste0('gene', 1:1000 )
colnames(network) <-paste0('gene', 1:1000 )
i <- 1
studies <- paste0('study', 1:10 )
nettype <- 'random'
dir <- 'out_test'
label <- studies[i]
filename <- 'output'


counts <- matrix(  2 ^ rpois(1e5, 5) , ncol=10, nrow=1000)
rownames(counts) <- paste0('gene', 1:1000 )
colnames(counts) <- paste0('sample', 1:10 )

groups <- c( rep(1,5), rep(2,5) )
method <- 'wilcox'

deg <- calc_DE(counts, groups, method)

subset_network(deg$degs, network)

}
