% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network_hdf5.R
\name{subset_network_hdf5}
\alias{subset_network_hdf5}
\title{Get sub-network}
\usage{
subset_network_hdf5(
  deg,
  network_type = "generic",
  method = "top",
  n_top = 100,
  min_cpm = 1,
  q = 0.05,
  fct = 2,
  flag_occr = TRUE,
  dir = ""
)
}
\arguments{
\item{deg}{A matrix (differential expression analysis)}

\item{network_type}{name of pre-built network to use (generic, blood, brain)}

\item{method}{top (fold change only) or reg (fold change and FDR cutoff)}

\item{n_top}{number of genes to consider}

\item{q}{q-value}

\item{fct}{abs fold change threshold}

\item{flag_occr}{boolean to select occurence network or coexpression network}

\item{dir}{dummy directory for the networks}
}
\value{
\code{combined_output}
}
\description{
Get sub-network
}
\examples{

i <- 1
studies <- paste0('study', 1:10 )
nettype <- 'random'
dir <- 'out_test'
label <- studies[i]
filename <- 'output'
genes <- rhdf5::h5read('/directflow/SCCGGroupShare/projects/sarba2/data/agg_coexp/generic.genes.h5',
 'genes' )


counts <- matrix(  2 ^ rpois(1e5, 5) , ncol=10, nrow=1000)
rownames(counts) <- sample(genes, 1000 )
colnames(counts) <- paste0('sample', 1:10 )

groups <- c( rep(1,5), rep(2,5) )
method <- 'wilcox'

deg <- calc_DE(counts, groups, method)
dir <- '/directflow/SCCGGroupShare/projects/sarba2/data/agg_coexp/'
dir <- 'S:/data/agg_coexp/'

network_type='generic'
subset_network_hdf5(deg$degs, network_type, dir=dir)

}
