% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network_hdf5.R
\name{subset_network_hdf5_gene_list}
\alias{subset_network_hdf5_gene_list}
\title{Get sub-network}
\usage{
subset_network_hdf5_gene_list(
  gene_list,
  network_type = "generic",
  flag_occr = TRUE,
  dir = ""
)
}
\arguments{
\item{gene_list}{A list of genes (entrez ID)}

\item{network_type}{name of pre-built network to use (generic, blood, brain)}

\item{flag_occr}{boolean to select occurence network or coexpression network}

\item{dir}{dummy directory for the networks}
}
\value{
\code{combined_output}
}
\description{
Get sub-network
}
\examples{

dir <- '/directflow/SCCGGroupShare/projects/sarba2/data/agg_coexp/'
dir <- 'S:/data/agg_coexp/'

network_type='generic'
subset_network_hdf5_gene_list(gene_list, network_type, dir=dir)

}
