% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_set_enrichment.R
\name{plot_gene_set_enrichment}
\alias{plot_gene_set_enrichment}
\title{Plot recurrence histogram}
\usage{
plot_gene_set_enrichment(data, gene_list, gene_sets)
}
\arguments{
\item{data}{enrichment results}

\item{gene_list}{genes}

\item{gene_sets}{gene sets}
}
\description{
Plot recurrence histogram
}
