% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot network}
\usage{
plot_network(sub_net, clust_net, threshold = 0.5, filt_min = 6)
}
\arguments{
\item{recur}{values}

\item{fdrs}{output from calc_fdrs_recur}

\item{n_max}{total number of studies}

\item{flag_plot}{plot type}
}
\description{
Plot network
}
\examples{

network <- diag(1000)
upper <- row(network) < col(network)
network[upper] <- runif(sum(upper), 0,1 )
network <- network + t(network)
diag(network) <- 1
rownames(network) <-paste0('gene', 1:1000 )
colnames(network) <-paste0('gene', 1:1000 )
i <- 1
studies <- paste0('study', 1:10 )
nettype <- 'random'
dir <- 'out_test'
label <- studies[i]
filename <- 'output'


cpm <- matrix(  abs(rnorm(10000)*10) , ncol=10, nrow=1000)
rownames(cpm) <- paste0('gene', 1:1000 )
colnames(cpm) <- studies

groups <- c( rep(1,5), rep(2,5) )
method <- 'wilcox'

deg <- calc_DE(cpm, groups, method)

subset_network(deg, network)

}
